+++ queue/src/main/resources/solr-doc-store-queue/migration/V1_1__Initial-schema.sql
@@ -1,11 +1,11 @@
 CREATE TABLE queue (
     agencyId NUMERIC(6) NOT NULL,
     bibliographicRecordId TEXT NOT NULL,
-    commitWithin BIGINT
+    commitWithin INTEGER
 );
 
 CREATE TABLE queue_error(
     agencyId NUMERIC(6) NOT NULL,
     bibliographicRecordId TEXT NOT NULL,
-    commitWithin BIGINT
+    commitWithin INTEGER
 );
